# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AttributeOrigin": {
        "additionalProperties": false,
        "properties": {
          "captured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_record_id": {
            "title": "Source Record Id",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          }
        },
        "required": [
          "source",
          "source_version",
          "source_record_id"
        ],
        "title": "AttributeOrigin",
        "type": "object"
      },
...
      "GameAttributes": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "developer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Developer"
          },
          "edition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "included_items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Items"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "release_date_granularity": {
            "anyOf": [
              {
                "enum": [
                  "year",
                  "month",
                  "day"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date Granularity"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "title": "GameAttributes",
        "type": "object"
      },
...
      "GameListResponse": {
        "properties": {
          "commerce_status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Commerce Status",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "commerce_status"
        ],
        "title": "GameListResponse",
        "type": "object"
      },
      "GameResponse": {
        "additionalProperties": false,
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/GameAttributes"
          },
          "commerce_status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Commerce Status",
            "type": "string"
          },
          "cover_attribution": {
            "title": "Cover Attribution",
            "type": "string"
          },
          "cover_url": {
            "title": "Cover Url",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "offers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OfferResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offers"
          },
          "origin_by_attribute": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AttributeOrigin"
            },
            "title": "Origin By Attribute",
            "type": "object"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_record_id": {
            "title": "Source Record Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "verified_at": {
            "title": "Verified At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "platform",
          "attributes",
          "origin_by_attribute",
          "source",
          "source_record_id",
          "version",
          "verified_at",
          "cover_attribution",
          "cover_url",
          "commerce_status"
        ],
        "title": "GameResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OfferResponse": {
        "additionalProperties": false,
        "properties": {
          "available_units": {
            "title": "Available Units",
            "type": "integer"
          },
          "condition_summary": {
            "title": "Condition Summary",
            "type": "string"
          },
          "currency": {
            "const": "BRL",
            "title": "Currency",
            "type": "string"
          },
          "demo_rank": {
            "title": "Demo Rank",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "enum": [
              "purchase",
              "rental"
            ],
            "title": "Mode",
            "type": "string"
          },
          "price_minor": {
            "title": "Price Minor",
            "type": "integer"
          },
          "sandbox": {
            "title": "Sandbox",
            "type": "boolean"
          },
          "sku_code": {
            "title": "Sku Code",
            "type": "string"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/PhysicalUnitResponse"
            },
            "title": "Units",
            "type": "array"
          }
        },
        "required": [
          "id",
          "sku_code",
          "mode",
          "price_minor",
          "currency",
          "condition_summary",
          "available_units",
          "demo_rank",
          "sandbox",
          "units"
        ],
        "title": "OfferResponse",
        "type": "object"
      },
      "PhysicalUnitResponse": {
        "additionalProperties": false,
        "properties": {
          "condition_summary": {
            "title": "Condition Summary",
            "type": "string"
          },
          "defects": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defects"
          },
          "included_items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Items"
          }
        },
        "required": [
          "condition_summary"
        ],
        "title": "PhysicalUnitResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RetroVault API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/games": {
      "get": {
        "operationId": "list_games_api_v1_catalog_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "query",
            "name": "genre",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Genre"
            }
          },
          {
            "in": "query",
            "name": "availability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "available",
                    "unavailable"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Availability"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "catalog",
              "enum": [
                "catalog",
                "demo_popular"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Commerce indispon\u00edvel."
          }
        },
        "summary": "List Games",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/games/{game_id}": {
      "get": {
        "operationId": "get_game_api_v1_catalog_games__game_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Validator for the published game.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "The published game has not changed.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Current game validator.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Published game not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Catalog unavailable."
          }
        },
        "summary": "Get Game",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/games/{game_id}/box-art": {
      "get": {
        "operationId": "get_box_art_api_v1_catalog_games__game_id__box_art_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Authorized published cover image.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Validator for the image bytes.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "The published cover has not changed.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Current image validator.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Published box art not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Box Art",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}